split models.dart
