+++ lib/utils/data_utils.dart
@@ -5,6 +5,21 @@ import 'package:convert/convert.dart';
 import 'package:crypto/crypto.dart' as crypto;
 import 'package:intl/intl.dart';
 
+typedef dynamic OnCall(List);
+
+class VarargsFunction extends Function {
+  OnCall _onCall;
+
+  VarargsFunction(this._onCall);
+
+  call() => _onCall([]);
+
+  noSuchMethod(Invocation invocation) {
+    final arguments = invocation.positionalArguments;
+    return _onCall(arguments);
+  }
+}
+
 int generateId([int length = 10]) {
   if (length < 4 || length > 18) throw Exception('Invalid length');
   String now = DateTime.now().millisecondsSinceEpoch.toString();
