+++ shared_prefs.db
@@ -1,9 +1,8 @@
 {"version":1,"sembast":1}
 {"key":"theme","value":"dark"}
 {"key":"homeLayout","value":"list"}
-{"key":"homeOptions","value":"{\"list\":{\"sorter\":\"creation\",\"sortRev\":true},\"calendar\":{}}"}
-{"key":"focusedDate","value":"1553784054135"}
-{"key":"homeLayout","value":"list"}
-{"key":"homeOptions","value":"{\"list\":{\"sorter\":\"creation\",\"sortRev\":true},\"calendar\":{}}"}
-{"key":"focusedDate","value":"1553784054135"}
-{"key":"homeOptions","value":"{\"list\":{\"sorter\":\"creation\",\"sortRev\":true},\"calendar\":{}}"}
+{"key":"homeOptions","value":"{\"list\":{\"sorter\":\"creation\",\"sortRev\":false},\"calendar\":{}}"}
+{"key":"focusedDate","value":"1552996800000"}
+{"key":"filters","value":"{\"type\":[]}"}
+{"key":"filters","value":"{\"type\":[]}"}
+{"key":"homeOptions","value":"{\"list\":{\"sorter\":\"creation\",\"sortRev\":false},\"calendar\":{}}"}
